Use non-abbreviated variable names
Although the proposal on Emmet-style abbreviated variable names -
https://www.mediawiki.org/wiki/Manual_talk:Coding_conventions/CSS/Archive_1#A_new_proposal -
wasn't strongly opposed, after talking to developers and community members
it seems to be the safe way to go with the full properties in order not to
put (possible) hurdles in the way. Might be revisited later.
Also fixing minor variable inheritance issue.